requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.13.1
gensim==4.2.0
mysqlclient==2.1.1
numpy==1.23.2
pytz==2022.2.1
scipy==1.9.1
smart-open==6.1.0
sqlparse==0.4.2
